Start using WooShippingAddPackageViewModel
