+++ WooCommerce/WooCommerceTests/ViewRelated/Shipping Label/WooShipping Create Shipping Labels/WooCarrierPackagesSelectionViewModelTests.swift
@@ -49,3 +49,172 @@ final class WooCarrierPackagesSelectionViewModelTests: XCTestCase {
         XCTAssertEqual(viewModel.selectedCarrierTab?.carrier.name, carrierTabs.last?.carrier.name)
     }
 }
+
+final class MockWooShippingPackagesRepository: WooShippingPackagesRepositoryProtocol {
+    private(set) var loadingSavedPackages: Bool = false
+    private(set) var customSavedPackages: [any WooShippingPackageDataRepresentable] = []
+    private(set) var predefinedSavedPackages: [any WooShippingPackageDataRepresentable] = []
+    private(set) var loadingCarrierPackages: Bool = false
+    @Published private(set) var carrierPackages: [WooShippingCarrierPackages] = []
+    var carrierPackagesPublisher: Published<[WooShippingCarrierPackages]>.Publisher { $carrierPackages }
+
+    func loadPackages() {
+        loadSavedPackages()
+        loadCarrierPackages()
+    }
+
+    func loadSavedPackages() {
+        loadingSavedPackages = true
+
+        customSavedPackages = [
+            WooShippingPackageData(id: UUID().uuidString,
+                                name: "Small Flat Rate Box",
+                                length: "21.92",
+                                width: "13.67",
+                                height: "4.14",
+                                dimensionsUnit: "cm",
+                                weight: "5",
+                                weightUnit: "kg",
+                                source: .custom,
+                                packageType: "box"),
+            WooShippingPackageData(id: UUID().uuidString,
+                                name: "Small Flat Rate Box",
+                                length: "21.92",
+                                width: "13.67",
+                                height: "4.14",
+                                dimensionsUnit: "cm",
+                                weight: "5",
+                                weightUnit: "kg",
+                                source: .custom,
+                                packageType: "box"),
+            WooShippingPackageData(id: UUID().uuidString,
+                                name: "Small Flat Rate Box",
+                                length: "21.92",
+                                width: "13.67",
+                                height: "4.14",
+                                dimensionsUnit: "cm",
+                                weight: "5",
+                                weightUnit: "kg",
+                                source: .custom,
+                                packageType: "box"),
+        ]
+        predefinedSavedPackages = [
+            WooShippingPackageData(name: "Small Flat Rate Box 2",
+                                  length: "21.92",
+                                  width: "13.67",
+                                  height: "4.14",
+                                  dimensionsUnit: "cm",
+                                  weight: "5",
+                                  weightUnit: "kg",
+                                  source: .predefined("DHL Express"),
+                                  packageType: "box"),
+            WooShippingPackageData(name: "Small Flat Rate Box 3",
+                                  length: "21.92",
+                                  width: "13.67",
+                                  height: "4.14",
+                                  dimensionsUnit: "cm",
+                                  weight: "5",
+                                  weightUnit: "kg",
+                                  source: .predefined("DHL Express"),
+                                  packageType: "box"),
+        ]
+
+        loadingSavedPackages = false
+    }
+
+    func loadCarrierPackages() {
+        loadingCarrierPackages = true
+
+        let uspsPackageGroups: [WooPackageGroup] = [
+            WooPackageGroup(name: "Flat Rate Boxes 1", packages: [
+                WooShippingPackageData(name: "Small Flat Rate Box 1",
+                                      length: "21.92",
+                                      width: "13.67",
+                                      height: "4.14",
+                                      dimensionsUnit: "cm",
+                                      weight: "5",
+                                      weightUnit: "kg",
+                                      source: .predefined("USPS Priority Mail Flat Rate Boxes"),
+                                      packageType: "box")
+            ]),
+            WooPackageGroup(name: "Flat Rate Boxes 2", packages: [
+                WooShippingPackageData(name: "Small Flat Rate Box 2",
+                                      length: "21.92",
+                                      width: "13.67",
+                                      height: "4.14",
+                                      dimensionsUnit: "cm",
+                                      weight: "5",
+                                      weightUnit: "kg",
+                                      source: .predefined("USPS Priority Mail Flat Rate Boxes"),
+                                      packageType: "box"),
+                WooShippingPackageData(name: "Small Flat Rate Box 21",
+                                      length: "21.92",
+                                      width: "13.67",
+                                      height: "4.14",
+                                      dimensionsUnit: "cm",
+                                      weight: "5",
+                                      weightUnit: "kg",
+                                      source: .predefined("USPS Priority Mail Flat Rate Boxes"),
+                                      packageType: "box"),
+                WooShippingPackageData(name: "Small Flat Rate Box 22",
+                                      length: "21.92",
+                                      width: "13.67",
+                                      height: "4.14",
+                                      dimensionsUnit: "cm",
+                                      weight: "5",
+                                      weightUnit: "kg",
+                                      source: .predefined("USPS Priority Mail Flat Rate Boxes"),
+                                      packageType: "box"),
+            ])
+        ]
+        let dhlPackageGroups: [WooPackageGroup] = [
+            WooPackageGroup(name: "Flat Rate Boxes 3", packages: [
+                WooShippingPackageData(name: "Small Flat Rate Box 3",
+                                      length: "21.92",
+                                      width: "13.67",
+                                      height: "4.14",
+                                      dimensionsUnit: "cm",
+                                      weight: "5",
+                                      weightUnit: "kg",
+                                      source: .predefined("DHL Express"),
+                                      packageType: "box"),
+            ]),
+            WooPackageGroup(name: "Flat Rate Boxes 4", packages: [
+                WooShippingPackageData(name: "Small Flat Rate Box 4",
+                                      length: "21.92",
+                                      width: "13.67",
+                                      height: "4.14",
+                                      dimensionsUnit: "cm",
+                                      weight: "5",
+                                      weightUnit: "kg",
+                                      source: .predefined("DHL Express"),
+                                      packageType: "box"),
+            ])
+        ]
+        let uspsCarrier: WooShippingCarrierPackages = WooShippingCarrierPackages(carrier: WooShippingCarrier.usps, packageGroups: uspsPackageGroups)
+        let dhlCarrier: WooShippingCarrierPackages = WooShippingCarrierPackages(carrier: WooShippingCarrier.dhlExpress, packageGroups: dhlPackageGroups)
+
+        carrierPackages = [uspsCarrier, dhlCarrier]
+
+        loadingCarrierPackages = false
+    }
+
+    // MARK: - Packages updates
+
+    func deleteSavedPackage(_ packageToRemove: WooShippingPackageDataRepresentable) async -> Error? {
+        customSavedPackages.removeAll { package in package.id == packageToRemove.id }
+        predefinedSavedPackages.removeAll { package in package.id == packageToRemove.id }
+
+        return nil
+    }
+
+    func addCustomPackage(_ packageToAdd: WooShippingPackageDataRepresentable) async -> Error? {
+        customSavedPackages.append(packageToAdd)
+        return nil
+    }
+
+    func addPredefinedPackage(_ packageToAdd: WooShippingPackageDataRepresentable) async -> Error? {
+        predefinedSavedPackages.append(packageToAdd)
+        return nil
+    }
+}
